General refactor of some apps
In order to enable more flexibility a lot of codes in threads,
comments, attachments, and notification should undergo
refactoring.

- Remove pre-save and post-save actions in the overidden save method
  in threads and comments models then push them to utility files.

This change makes it possible for us to call save() for any form of
update instead coupling it to only full post create and update.

- Remove all the methods for creating notification for all the
  notification types in NotificationQuerySet.

This methods can be considered duplication of the event types in the
Notification model. Instead, we can easily use a single method to
perform all this actions. It should be duty of clients to the
Notification model to prepare the kind of notification want the single
method to create.